split up the flow
