+++ src/pages/flow/index.js
@@ -1,45 +1,11 @@
 import React from 'react';
-import { map } from 'microstates';
 import { State } from '../../state';
 import BarChart from './barChart';
 
-import TabView from '../../components/view/tabView';
-import TransactionInput from '../../forms/transactionInput';
-import AccountInput from '../../forms/accountInput';
-import AccountTransactionInput from '../../forms/accountTransactionInput';
+import Transactions from './transactions';
+import Accounts from './accounts';
 
 class Financial extends React.Component {
-  constructor(props) {
-    super();
-    this.setAccountForm = this.setAccountForm.bind(this);
-    this.setTransactionForm = this.setTransactionForm.bind(this);
-    this.tabClickAccounts = this.tabClickAccounts.bind(this);
-    this.tabClickTransactions = this.tabClickTransactions.bind(this);
-    this.state = {
-      activeTabTransactions: 0,
-      activeTabAccounts: 0,
-      TransactionType: { income: true, expense: true, transfer: true }
-    };
-  }
-
-  tabClickTransactions(index) {
-    this.setState({ activeTabTransactions: index });
-  }
-
-  setTransactionForm(model, index, id) {
-    this.setState({ activeTabTransactions: index });
-    model.modifyTransaction(id);
-  }
-
-  tabClickAccounts(index) {
-    this.setState({ activeTabAccounts: index });
-  }
-
-  setAccountForm(model, index, name) {
-    this.setState({ activeTabAccounts: index });
-    model.modifyAccount(name);
-  }
-
   render() {
     return (
       <State.Consumer>
@@ -82,114 +48,8 @@ class Financial extends React.Component {
               </div>
               <BarChart data={model.charts.state} />
             </section>
-
-            <section className="section" id="transactions">
-              <TabView
-                activeTab={this.state.activeTabTransactions}
-                tabClick={this.tabClickTransactions}
-                tabTitles={[
-                  'All Transactions',
-                  'Add Transaction',
-                  'Income',
-                  'Expenses',
-                  'Transfers'
-                ]}
-                tabContents={[
-                  <React.Fragment>
-                    <div className="buttons">
-                      {Object.keys(model.state.transactionCategories).map(
-                        category => (
-                          <button
-                            key={category}
-                            className={
-                              model.state.transactionCategories[category]
-                                ? 'button is-primary'
-                                : 'button is-secondary'
-                            }
-                            onClick={model.filterTransactionsComputed.bind(
-                              this,
-                              category
-                            )}
-                          >
-                            {category}
-                          </button>
-                        )
-                      )}
-                    </div>
-                    {transactionTable(model.state.transactionsComputed, {
-                      model: model,
-                      setTransactionForm: this.setTransactionForm,
-                      deleteTransaction: model.deleteTransaction
-                    })}
-                  </React.Fragment>,
-                  <TransactionInput tabClick={this.tabClickTransactions} />,
-                  transactionTable(
-                    model.state.transactionsComputed.filter(
-                      transaction => transaction.type === 'income'
-                    ),
-                    {
-                      model: model,
-                      setTransactionForm: this.setTransactionForm,
-                      deleteTransaction: model.deleteTransaction
-                    }
-                  ),
-                  transactionTable(
-                    model.state.transactionsComputed.filter(
-                      transaction => transaction.type === 'expense'
-                    ),
-                    {
-                      model: model,
-                      setTransactionForm: this.setTransactionForm,
-                      deleteTransaction: model.deleteTransaction
-                    }
-                  ),
-                  transactionTable(
-                    model.state.transactionsComputed.filter(
-                      transaction => transaction.type === 'transfer'
-                    ),
-                    {
-                      model: model,
-                      setTransactionForm: this.setTransactionForm,
-                      deleteTransaction: model.deleteTransaction
-                    }
-                  )
-                ]}
-              />
-            </section>
-
-            <section className="section" id="accounts">
-              <TabView
-                activeTab={this.state.activeTabAccounts}
-                tabClick={this.tabClickAccounts}
-                tabTitles={['All Accounts', 'Add Account', 'Debt']}
-                tabContents={[
-                  accountTable(model.accountsComputed, {
-                    model: model,
-                    setAccountForm: this.setAccountForm,
-                    deleteAccount: model.deleteAccount,
-                    toggleAccountVisibility: model.toggleAccountVisibility
-                  }),
-                  <AccountInput tabClick={this.tabClickAccounts} />,
-                  <React.Fragment>
-                    <div>
-                      {debtTable(model.state.accountsComputed, {
-                        model: model,
-                        setAccountForm: this.setAccountForm
-                      })}
-                    </div>
-                    <div>
-                      {model.state.accountsComputed.filter(
-                        account => account.vehicle === 'debt'
-                      ).length === 0 ? null : (
-                        <AccountTransactionInput
-                          tabClick={this.tabClickAccounts}
-                        />
-                      )}
-                    </div>
-                  </React.Fragment>
-                ]}
-              />
-            </section>
+            <Transactions />
+            <Accounts />
           </React.Fragment>
         )}
       </State.Consumer>
@@ -198,191 +58,3 @@ class Financial extends React.Component {
 }
 
 export default Financial;
-
-const transactionTable = (data, actions) =>
-  data.length === 0 || !data ? (
-    <div>There are no transactions to show.</div>
-  ) : (
-    <table className="table is-striped is-hoverable">
-      <thead>
-        <tr>
-          <th>
-            <abbr title="real account">raccount</abbr>
-          </th>
-          <th>description</th>
-          <th>category</th>
-          <th>type</th>
-          <th>
-            <abbr title="start date">start</abbr>
-          </th>
-          <th>
-            <abbr title="repeat type">rtype</abbr>
-          </th>
-          <th>cycle</th>
-          <th>value</th>
-          <th>Daily Rate</th>
-          <th>Modify</th>
-          <th>Delete</th>
-        </tr>
-      </thead>
-      <tbody>
-        {map(data.filter(state => !state.fromAccount), transaction => (
-          <tr key={transaction.id}>
-            <td>{transaction.raccount}</td>
-            <td>{transaction.description}</td>
-            <td>{transaction.category}</td>
-            <td>{transaction.type}</td>
-            <td>{transaction.start}</td>
-            <td>{transaction.rtype}</td>
-            <td>{!transaction.cycle ? '' : transaction.cycle.toFixed(0)}</td>
-            <td>{!transaction.value ? '' : transaction.value.toFixed(2)}</td>
-            <td>{transaction.dailyRate.toFixed(2)}</td>
-            <td>
-              <button
-                className="button is-rounded is-small is-info"
-                onClick={() =>
-                  actions.setTransactionForm(actions.model, 1, transaction.id)
-                }
-              >
-                M
-              </button>
-            </td>
-            <td>
-              <button
-                className="delete"
-                onClick={actions.deleteTransaction.bind(this, transaction.id)}
-              />
-            </td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
-
-const accountTable = (data, actions) =>
-  data.length === 0 || !data ? (
-    <div>There are no accounts to show.</div>
-  ) : (
-    <table className="table is-striped is-hoverable">
-      <thead>
-        <tr>
-          <th />
-          <th>name</th>
-          <th>
-            <abbr title="starting balance">starting</abbr>
-          </th>
-          <th>interest</th>
-          <th>vehicle</th>
-          <th>Modify</th>
-          <th>Delete</th>
-        </tr>
-      </thead>
-      <tbody>
-        {map(data, account => (
-          <tr key={account.name.state}>
-            <td
-              onClick={actions.toggleAccountVisibility.bind(
-                this,
-                account.name.state
-              )}
-            >
-              {account.visible.state ? `👀` : `🤫`}
-            </td>
-            <th>{account.name.state}</th>
-            <td>{account.starting.toFixed}</td>
-            <td>{account.interest.toFixed}%</td>
-            <td>{account.vehicle.state}</td>
-            <td>
-              <button
-                className="button is-rounded is-small is-info"
-                onClick={() =>
-                  actions.setAccountForm(actions.model, 1, account.name.state)
-                }
-              >
-                M
-              </button>
-            </td>
-            <td>
-              <button
-                className="delete"
-                onClick={actions.deleteAccount.bind(this, account.name.state)}
-              />
-            </td>
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
-
-const debtTable = (data, actions) =>
-  data.filter(account => account.vehicle === 'debt').length === 0 || !data ? (
-    <div>There are no debts to show.</div>
-  ) : (
-    data
-      .filter(account => account.vehicle === 'debt')
-      .map(account => (
-        <div className="media box" key={account.name}>
-          <div className="media-content">
-            <div className="content">
-              <p>
-                <strong>{account.name}</strong>{' '}
-                <small>{`$${account.starting} @ ${account.interest}%`}</small>
-              </p>
-            </div>
-            {account.payback ? paybackTable(account, actions) : null}
-          </div>
-          <div className="media-right">
-            <button
-              className="button is-rounded is-small is-success"
-              onClick={actions.model.forms.testing}
-            >
-              +
-            </button>
-            <button
-              className="button is-rounded is-small is-info"
-              onClick={() =>
-                actions.setAccountForm(actions.model, 1, account.name)
-              }
-            >
-              M
-            </button>
-            <button
-              className="delete"
-              onClick={() => actions.model.deleteAccount(account.name)}
-            />
-          </div>
-        </div>
-      ))
-  );
-
-const paybackTable = (data, actions) =>
-  data.payback.transactions.map((paybackTransaction, index) => (
-    <div className="media" key={index}>
-      <div className="media-content">
-        <p>
-          <strong>{paybackTransaction.start}</strong>{' '}
-          <small>{`${paybackTransaction.rtype} @ ${
-            paybackTransaction.cycle
-          } for ${paybackTransaction.value}`}</small>
-        </p>
-        <p>{paybackTransaction.description}</p>
-      </div>
-      <div className="media-right">
-        <button
-          className="button is-rounded is-small is-info"
-          onClick={() => {
-            this.toggleAccountTransactionVisibility();
-            actions.model.modifyAccountTransaction(data.name, index);
-          }}
-        >
-          M
-        </button>
-        <button
-          className="delete"
-          onClick={() =>
-            actions.model.modifyAccountTransaction(data.name, index)
-          }
-        />
-      </div>
-    </div>
-  ));
